DEX-932 Starting refactoring
* The Matcher class was merged with the Application;
